Added Upsert unit test and fix relocated indexing
